Store whole creature state in querystring.
